+++ src/stats/serialize.rs
@@ -3,8 +3,8 @@ use super::*;
 #[derive(Serialize, Deserialize, Debug)]
 pub struct ItemStatsSerializer {
     pub item: String,
-    pub frecency: f32,
-    pub last_accessed: f32,
+    pub frecency: f64,
+    pub last_accessed: f64,
     pub num_accesses: i32,
 }
 
@@ -20,7 +20,7 @@ impl From<ItemStats> for ItemStatsSerializer {
 }
 
 impl ItemStatsSerializer {
-    pub fn into_item_stats(self, ref_time: f64, half_life: f32) -> ItemStats {
+    pub fn into_item_stats(self, ref_time: f64, half_life: f64) -> ItemStats {
         ItemStats {
             half_life,
             reference_time: ref_time,
